+++ Frontend/src/pages/QuestionFormEdit.jsx
@@ -4,41 +4,48 @@ import { Link, useNavigate} from "react-router-dom";
 import { useLocation } from "react-router-dom";
 import {React, useEffect } from 'react';
 import Question from '../components/QuestionCard';
+import $ from "jquery";
 
 
 
 const QuestionFormEdit = (id ) => {
 
+    var dropdownV="ScId";
     const location = useLocation();
     const idQuestion= location.state.id
 
     const url= "http://localhost:8000/api/question/"+idQuestion+"/";
    
     const [questions, setQuiz] = useState([])
-    const navigate = useNavigate();
     const [questionText, setQuestionText] = useState('')
     const [defaultAnswer, setDefaultAnswer] = useState('')
     const [author, setAuthorId] = useState('')
     const [multiplayer, setMutliplayer] = useState('false')
     const [questiontype, setQuestionType] = useState('MC')
     const [questionAnswerOption, setQuestionAnswerOption] = useState('')
+
+    const $ = require( "jquery" );
+    const navigate = useNavigate();
     
     function deleteItem(event){
         event.preventDefault()
         console.log(idQuestion)
-        
-        axios(
-            {
-                method: "DELETE",
-                url: url,
-                headers: {'Content-Type': 'application/json'}
-            }
-        ).then((response) => {
-            console.log(response.data)
-        })
+        if (window.confirm('Do you really want to delete this question?')){
+            axios(
+                {
+                    method: "DELETE",
+                    url: url,
+                    headers: {'Content-Type': 'application/json'}
+                }
+            ).then((response) => {
+                console.log(response.data)
+            })
 
-        navigate("/Library", 
-        )
+            navigate("/Library", 
+            )
+        }else{
+            // They clicked no
+        }
         
     }
 
@@ -49,8 +56,18 @@ const QuestionFormEdit = (id ) => {
             console.log(data)
             setQuiz(data)
             setQuestionText(data.question_text)
-            setDefaultAnswer(data.defaultAnswer)
+            setDefaultAnswer(data.defaultAnswer)//nested didnt work
             setAuthorId(1)
+            setQuestionType(data.question_type)
+            setMutliplayer(data.multiplayer)
+            setQuestionAnswerOption(data.question_answer_option)//array von objecten works
+            console.log(data.question_text)
+            console.log(data.defaultAnswer)
+            console.log(1)
+            console.log(data.question_type)
+            console.log(data.multiplayer)
+            console.log(data.question_answer_option)
+
         }
         else {
             console.log(response.status)
@@ -127,6 +144,28 @@ const QuestionFormEdit = (id ) => {
         
     
     }
+
+    function changeQuestion(value){
+        console.log(value)
+        
+        var mcString ='<div id= "containerID2" className="container2"> <label htmlFor="exampleFormControlInput2">Choice 2</label><div><input type="text" class="form-control" id="exampleFormControlInput2" placeholder="New Answer" text={defaultAnswer} ></input><input className="right" type="checkbox"></input> <label htmlFor="exampleFormControlInput1">True</label></div></div>'
+        var mcString2='<div id= "containerID3" className="container3"> <label htmlFor="exampleFormControlInput3">Choice 3</label><div><input type="text" class="form-control" id="exampleFormControlInput3" placeholder="New Answer" text={defaultAnswer} ></input><input className="right" type="checkbox"></input> <label htmlFor="exampleFormControlInput1">True</label></div></div>'
+        var mcString3='<div id= "containerID4" className="container4"> <label htmlFor="exampleFormControlInput4">Choice 4</label><div><input type="text" class="form-control" id="exampleFormControlInput4" placeholder="New Answer" text={defaultAnswer} ></input><input className="right" type="checkbox"></input> <label htmlFor="exampleFormControlInput1">True</label></div></div>'
+        if(value==="MC"){
+            $(".container1").append(mcString+mcString2+mcString3)
+            dropdownV="McId"
+        }else{
+            if(dropdownV==="McId"){
+                const element2 = document.getElementById('containerID2');
+                const element3 = document.getElementById('containerID3');
+                const element4 = document.getElementById('containerID4');
+                element2.remove()
+                element3.remove()
+                element4.remove()
+                dropdownV=value
+            }
+        }
+    }
     
     return (
         <>
@@ -138,10 +177,10 @@ const QuestionFormEdit = (id ) => {
                 <div className="custom-card col-lg-6 col-md-8 p-5 bg-dark justify-content-center align-self-center">
                 <form className="text-light" >
                         <label for="type">Choose a Type: </label>
-                        <select  id="type" name="type">
-                            <option value="SC">Single Choice</option>
-                            <option value="MC">Multiple Choice</option>
-                            <option value="EQ">Estimate Question</option>
+                        <select  id="selectOpt" name="typeSelection" onChange={(e) => changeQuestion(e.target.value)}>
+                            <option id= "ScId"value="SC">Single Choice</option>
+                            <option id="McId" value="MC">Multiple Choice</option>
+                            <option id= "EqId" value="EQ">Estimate Question</option>
                         </select>
                         <label className="mb-2 rechts-oben"  htmlFor="exampleFormControlInput1">Multiplayer </label> 
                         <input type="checkbox"/>
@@ -151,7 +190,7 @@ const QuestionFormEdit = (id ) => {
 
                         <label className="mb-2"  htmlFor="exampleFormControlInput1">Question Text</label>
                         <input type="text" class="form-control" id="exampleFormControlInput1"
-                            placeholder="New Question"
+                            placeholder={questionText}
                             text={questionText}
                             onChange={(e) => setQuestionText(e.target.value)}
                             ></input>
@@ -161,35 +200,8 @@ const QuestionFormEdit = (id ) => {
                         <div className="container1"> 
                             <label htmlFor="exampleFormControlInput1">Choice 1 (has to be true)</label>
                             <div>
-                            <input type="text" class="form-control" id="exampleFormControlInput1" placeholder="New Answer" text={defaultAnswer} 
+                            <input type="text" class="form-control" id="exampleFormControlInput1" placeholder={defaultAnswer} text={defaultAnswer} 
                             onChange={(e) => setDefaultAnswer(e.target.value)}></input>
-                            </div>
-                        </div>
-                        <div className="container2"> 
-                            <label htmlFor="exampleFormControlInput2">Choice 2</label>
-                            <div>
-                            <input type="text" class="form-control" id="exampleFormControlInput2" placeholder="New Answer" text={defaultAnswer} 
-                            onChange={(e) => setQuestionAnswerOption(questionAnswerOption.push(e.target.value))}
-                            ></input>
-                            <input className="right" type="checkbox"></input> <label htmlFor="exampleFormControlInput1">True</label>
-                            </div>
-                        </div>
-                        <div className="container3"> 
-                            <label htmlFor="exampleFormControlInput3">Choice 3</label>
-                            <div>
-                            <input type="text" class="form-control" id="exampleFormControlInput3" placeholder="New Answer" text={defaultAnswer} 
-                            onChange={(e) => setQuestionAnswerOption(e.target.value)}
-                            ></input>
-                            <input className="right" type="checkbox"></input> <label htmlFor="exampleFormControlInput1">True</label>
-                            </div>
-                        </div>
-                        <div className="container4"> 
-                            <label htmlFor="exampleFormControlInput4">Choice 4</label>
-                            <div>
-                            <input className="form-control" type="text"  id="exampleFormControlInput4" placeholder="New Answer" text={defaultAnswer} 
-                            onChange={(e) => setQuestionAnswerOption(e.target.value)}
-                            ></input>
-                            <input className="right" type="checkbox"></input> <label htmlFor="exampleFormControlInput1">True</label>
                             </div>
                         </div>
                     </form>
